split epoch operations into multiple files
